+++ nodes/Watsonx/LmChatWatsonX.node.ts
@@ -8,6 +8,7 @@ import {
 	INodePropertyOptions,
 } from 'n8n-workflow';
 import { N8nLlmTracing } from './depedancies/N8nLlmTracing';
+import { N8nLangfuseHandler } from './depedancies/N8nLangfuseHandler';
 import { makeN8nLlmFailedAttemptHandler } from './depedancies/n8nLlmFailedAttemptHandler';
 import { ChatWatsonx } from '@langchain/community/chat_models/ibm';
 
@@ -156,15 +157,34 @@ export class LmChatWatsonX implements INodeType {
 		const credentials = await this.getCredentials('watsonxApi', itemIndex);
 		const modelId = this.getNodeParameter('modelId', itemIndex) as string;
 		const options = this.getNodeParameter('options', itemIndex, {}) as IWatsonxOptions;
+		const callbacks: any[] = [new N8nLlmTracing(this)];
 
-
+		//Langfuse Handler Creation
+		try {
+			this.logger.debug("[WatsonX-Langfuse] Initializing Langfuse Callback Handler...");
+			const langfuse = new N8nLangfuseHandler({
+				publicKey: process.env.LANGFUSE_PUBLIC_KEY,
+				secretKey: process.env.LANGFUSE_SECRET_KEY,
+				baseUrl: process.env.LANGFUSE_HOST,
+				metadata: { 
+					modelName: modelId,
+					workflowName: this.getWorkflow().name as string
+				},
+				logger: this.logger,
+			});
+			callbacks.push(langfuse);
+			this.logger.debug('[WatsonX-Langfuse] Langfuse handler initialized, added to callbacks.');
+		} catch (error) {
+			this.logger.debug('[WatsonX-Langfuse] Langfuse credentials not found or invalid, skipping Langfuse tracing.');
+		}
+		
 		const props: any = {
 			model: modelId,
 			version: this.getNodeParameter('version', itemIndex),
 			projectId: credentials.projectId,
 			stream: false,
 			...options,
-			callbacks: [new N8nLlmTracing(this)],
+			callbacks,
 			onFailedAttempt: makeN8nLlmFailedAttemptHandler(this),
 		};
